requirements.txt
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
enum34==1.1.6
gast==0.2.2
google-pasta==0.1.7
graphviz==0.10.1
grpcio==1.19.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.4
mock==2.0.0
more-itertools==7.0.0
numpy==1.16.2
opencv-python==4.0.1.24
pandas==0.22.0
pbr==5.1.3
Pillow==5.4.1
plaidml==0.5.0
protobuf==3.7.1
pydot==1.0.29
pydot-ng==2.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.11.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
Werkzeug==0.15.2
wrapt==1.11.2
xgboost==0.90
